# live_life/api_models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml

from __future__ import annotations

from datetime import date as date_aliased
from enum import StrEnum
from typing import Any

from pydantic import (
    AwareDatetime,
    BaseModel,
    ConfigDict,
    Field,
    RootModel,
    confloat,
    conint,
    constr,
)


class DateRange(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    from_: date_aliased = Field(..., alias='from')
    to: date_aliased


class SourceName(StrEnum):
    bracelet = 'bracelet'
    welltory = 'welltory'
    todoist = 'todoist'
    rescuetime = 'rescuetime'


class SourceRunStatus(StrEnum):
    success = 'success'
    partial = 'partial'
    failed = 'failed'
    not_run = 'not_run'


class QualityStatus(StrEnum):
    complete = 'complete'
    partial = 'partial'
    failed = 'failed'
    not_run = 'not_run'
    in_progress = 'in_progress'


class SourceStatus(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    source: SourceName
    status: SourceRunStatus
    lastRunAt: AwareDatetime | None = None


class MetricPoint(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    timestamp: AwareDatetime
    metric: str
    value: float
    unit: str | None = None


class TaskItem(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    content: str
    timestamp: AwareDatetime


class Perspective(StrEnum):
    activity = 'activity'
    productivity = 'productivity'


class RescueTimeItem(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    timestamp: AwareDatetime
    perspective: Perspective
    label: str
    seconds: confloat(ge=0.0)


class BraceletSummary(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    sleepSeconds: float | None
    steps: float | None


class TodoistSummary(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    created: conint(ge=0)
    completed: conint(ge=0)


class SourceIndicator(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    available: bool
    count: conint(ge=0)


class DayDetail(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    braceletMetrics: list[MetricPoint]
    welltoryMetrics: list[MetricPoint]
    createdTasks: list[TaskItem]
    completedTasks: list[TaskItem]
    rescueTime: list[RescueTimeItem]


class DashboardDay(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    date: date_aliased
    weekday: conint(ge=1, le=7)
    currentDay: bool
    quality: QualityStatus
    sources: list[SourceStatus]
    bracelet: BraceletSummary
    welltory: SourceIndicator
    todoist: TodoistSummary
    rescuetime: SourceIndicator
    detail: DayDetail


class DashboardResponse(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    from_: date_aliased = Field(..., alias='from')
    to: date_aliased
    timezone: str
    generatedAt: AwareDatetime
    days: list[DashboardDay]


class Action(StrEnum):
    reconnect = 'reconnect'
    configure = 'configure'
    retry = 'retry'


class SourceIssue(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    code: str
    message: str
    action: Action
    steps: list[str]
    diagnosticId: str | None = None


class Status(StrEnum):
    idle = 'idle'
    pending = 'pending'
    success = 'success'
    failed = 'failed'


class DriveConnectionState(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    status: Status
    sessionId: str | None = None
    authorizationUrl: str | None = None
    issue: SourceIssue | None = None


class DriveClientUpload(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    content: constr(max_length=16384)


class SourceSyncSummary(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    issue: SourceIssue | None = None
    source: SourceName
    status: SourceRunStatus
    records: conint(ge=0)


class DashboardSyncResponse(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    from_: date_aliased = Field(..., alias='from')
    to: date_aliased
    sources: list[SourceSyncSummary] = Field(..., max_length=4, min_length=4)


class ApiError(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    code: str
    message: str
    details: dict[str, Any]


class Type(StrEnum):
    progress = 'progress'


class SyncProgressEvent(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    issue: SourceIssue | None = None
    type: Type
    source: SourceName
    status: SourceRunStatus
    records: conint(ge=0)
    latest: AwareDatetime | None = None
    display: str | None = None


class Type1(StrEnum):
    complete = 'complete'


class SyncCompleteEvent(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    type: Type1
    from_: date_aliased = Field(..., alias='from')
    to: date_aliased
    sources: list[SourceSyncSummary] = Field(..., max_length=4, min_length=4)


class Type2(StrEnum):
    error = 'error'


class SyncErrorEvent(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    type: Type2
    code: str
    message: str


class SyncStreamEvent(
    RootModel[SyncProgressEvent | SyncCompleteEvent | SyncErrorEvent]
):
    root: SyncProgressEvent | SyncCompleteEvent | SyncErrorEvent