# openapi.yaml
openapi: 3.1.1
info:
  title: Live Life Local Dashboard API
  version: 1.0.0
  description: Local-only API for personal telemetry visualization and source synchronization.
servers:
  - url: http://127.0.0.1:8000
paths:
  /api/dashboard:
    get:
      operationId: getDashboard
      summary: Read dashboard data for an inclusive date range
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/data-sync:
    post:
      operationId: syncDashboardData
      summary: Synchronize every data source shown on the dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateRange'
      responses:
        '200':
          description: Source synchronization attempts completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSyncResponse'
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Another dashboard synchronization is running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/google-drive/connect:
    post:
      operationId: connectGoogleDrive
      summary: Start user-initiated Google Drive authorization
      responses:
        '200':
          description: Connection session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriveConnectionState'
        '403':
          $ref: '#/components/responses/BadRequest'
  /api/google-drive/connection/{sessionId}:
    get:
      operationId: getGoogleDriveConnection
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connection progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriveConnectionState'
        '403':
          $ref: '#/components/responses/BadRequest'
  /api/google-drive/client:
    post:
      operationId: saveGoogleDriveClient
      summary: Save a user-selected Desktop OAuth client locally
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriveClientUpload'
      responses:
        '200':
          description: Configuration saved or actionable failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriveConnectionState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/BadRequest'
components:
  parameters:
    FromDate:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date
    ToDate:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalError:
      description: The local operation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    DateRange:
      type: object
      additionalProperties: false
      required: [from, to]
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
    SourceName:
      type: string
      enum: [bracelet, welltory, todoist, rescuetime]
    SourceRunStatus:
      type: string
      enum: [success, partial, failed, not_run]
    QualityStatus:
      type: string
      enum: [complete, partial, failed, not_run, in_progress]
    SourceStatus:
      type: object
      additionalProperties: false
      required: [source, status]
      properties:
        source:
          $ref: '#/components/schemas/SourceName'
        status:
          $ref: '#/components/schemas/SourceRunStatus'
        lastRunAt:
          type: [string, 'null']
          format: date-time
    MetricPoint:
      type: object
      additionalProperties: false
      required: [timestamp, metric, value]
      properties:
        timestamp:
          type: string
          format: date-time
        metric:
          type: string
        value:
          type: number
        unit:
          type: [string, 'null']
    TaskItem:
      type: object
      additionalProperties: false
      required: [content, timestamp]
      properties:
        content:
          type: string
        timestamp:
          type: string
          format: date-time
    RescueTimeItem:
      type: object
      additionalProperties: false
      required: [timestamp, perspective, label, seconds]
      properties:
        timestamp:
          type: string
          format: date-time
        perspective:
          type: string
          enum: [activity, productivity]
        label:
          type: string
        seconds:
          type: number
          minimum: 0
    BraceletSummary:
      type: object
      additionalProperties: false
      required: [sleepSeconds, steps]
      properties:
        sleepSeconds:
          type: [number, 'null']
        steps:
          type: [number, 'null']
    TodoistSummary:
      type: object
      additionalProperties: false
      required: [created, completed]
      properties:
        created:
          type: integer
          minimum: 0
        completed:
          type: integer
          minimum: 0
    SourceIndicator:
      type: object
      additionalProperties: false
      required: [available, count]
      properties:
        available:
          type: boolean
        count:
          type: integer
          minimum: 0
    DayDetail:
      type: object
      additionalProperties: false
      required: [braceletMetrics, welltoryMetrics, createdTasks, completedTasks, rescueTime]
      properties:
        braceletMetrics:
          type: array
          items:
            $ref: '#/components/schemas/MetricPoint'
        welltoryMetrics:
          type: array
          items:
            $ref: '#/components/schemas/MetricPoint'
        createdTasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskItem'
        completedTasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskItem'
        rescueTime:
          type: array
          items:
            $ref: '#/components/schemas/RescueTimeItem'
    DashboardDay:
      type: object
      additionalProperties: false
      required:
        - date
        - weekday
        - currentDay
        - quality
        - sources
        - bracelet
        - welltory
        - todoist
        - rescuetime
        - detail
      properties:
        date:
          type: string
          format: date
        weekday:
          type: integer
          minimum: 1
          maximum: 7
        currentDay:
          type: boolean
        quality:
          $ref: '#/components/schemas/QualityStatus'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceStatus'
        bracelet:
          $ref: '#/components/schemas/BraceletSummary'
        welltory:
          $ref: '#/components/schemas/SourceIndicator'
        todoist:
          $ref: '#/components/schemas/TodoistSummary'
        rescuetime:
          $ref: '#/components/schemas/SourceIndicator'
        detail:
          $ref: '#/components/schemas/DayDetail'
    DashboardResponse:
      type: object
      additionalProperties: false
      required: [from, to, timezone, generatedAt, days]
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        timezone:
          type: string
        generatedAt:
          type: string
          format: date-time
        days:
          type: array
          items:
            $ref: '#/components/schemas/DashboardDay'
    SourceIssue:
      type: object
      additionalProperties: false
      required: [code, message, action, steps]
      properties:
        code:
          type: string
        message:
          type: string
        action:
          type: string
          enum: [reconnect, configure, retry]
        steps:
          type: array
          items:
            type: string
        diagnosticId:
          type: string
    DriveConnectionState:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [idle, pending, success, failed]
        sessionId:
          type: string
        authorizationUrl:
          type: string
        issue:
          $ref: '#/components/schemas/SourceIssue'
    DriveClientUpload:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content:
          type: string
          maxLength: 16384
    SourceSyncSummary:
      type: object
      additionalProperties: false
      required: [source, status, records]
      properties:
        issue:
          $ref: '#/components/schemas/SourceIssue'
        source:
          $ref: '#/components/schemas/SourceName'
        status:
          $ref: '#/components/schemas/SourceRunStatus'
        records:
          type: integer
          minimum: 0
    DashboardSyncResponse:
      type: object
      additionalProperties: false
      required: [from, to, sources]
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        sources:
          type: array
          minItems: 4
          maxItems: 4
          items:
            $ref: '#/components/schemas/SourceSyncSummary'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    SyncProgressEvent:
      type: object
      additionalProperties: false
      required: [type, source, status, records]
      properties:
        issue:
          $ref: '#/components/schemas/SourceIssue'
        type:
          type: string
          enum: [progress]
        source:
          $ref: '#/components/schemas/SourceName'
        status:
          $ref: '#/components/schemas/SourceRunStatus'
        records:
          type: integer
          minimum: 0
        latest:
          type: [string, 'null']
          format: date-time
        display:
          type: [string, 'null']
    SyncCompleteEvent:
      type: object
      additionalProperties: false
      required: [type, from, to, sources]
      properties:
        type:
          type: string
          enum: [complete]
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        sources:
          type: array
          minItems: 4
          maxItems: 4
          items:
            $ref: '#/components/schemas/SourceSyncSummary'
    SyncErrorEvent:
      type: object
      additionalProperties: false
      required: [type, code, message]
      properties:
        type:
          type: string
          enum: [error]
        code:
          type: string
        message:
          type: string
    SyncStreamEvent:
      oneOf:
        - $ref: '#/components/schemas/SyncProgressEvent'
        - $ref: '#/components/schemas/SyncCompleteEvent'
        - $ref: '#/components/schemas/SyncErrorEvent'